Minor ui modifications for dashboard, tv details screens

+++ features/tv/src/commonMain/kotlin/data/datasources/remote/models/GetTvReviewsResponse.kt
@@ -26,7 +26,7 @@ data class GetTvReviewsResponse(
         data class AuthorDetails(
             val avatar_path: String? = null,
             val name: String? = null,
-            val rating: Int? = null,
+            val rating: Double? = null,
             val username: String? = null
         )
     }
